// frontend/felcn-base-frontend/src/app/(fase_2)/apd/registro_caso/types/registro-caso-apd.types.ts
export interface UltimaSituacionJuridica {
  situacionId: string
  detenidoId: string
  situacionLegalId: string
  fecha?: string
  situacionLegal: {
    slId: number
    descripcion: string
  }
}
// ...
export interface PersonaImplicadaRow {
  deId: number
  casoId: number
  nombres: string
  paterno: string
  materno: string
  esposo: string
  paisId: number
  estadoCivilId: number
  profesionId: number
  tipoDocumentoId: number
  numeroDocumento: string
  relacion: string
  observaciones: string
  estado: boolean
  fechahoraing: string
  ultimaSituacionJuridica: UltimaSituacionJuridica | null
  [key: string]: unknown
}